Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/cors/status-async-expected.txt

CONSOLE MESSAGE: Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/status.py?code=699&headers=x-nonsimple&text=OHAI due to access control checks.
CONSOLE MESSAGE: Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/status.py?code=501&headers=x-nonsimple&text=OHAI due to access control checks.
CONSOLE MESSAGE: Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/status.py?code=401&headers=x-nonsimple&text=OHAI due to access control checks.
CONSOLE MESSAGE: Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/status.py?code=400&headers=x-nonsimple&text=OHAI due to access control checks.
CONSOLE MESSAGE: Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/status.py?code=404&headers=x-nonsimple&text=OHAI due to access control checks.
CONSOLE MESSAGE: Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/status.py?code=501&headers=x-nonsimple&text=OHAI due to access control checks.
CONSOLE MESSAGE: Preflight response is not successful
CONSOLE MESSAGE: XMLHttpRequest cannot load http://127.0.0.1:8800/cors/resources/status.py?code=699&headers=x-nonsimple&text=OHAI due to access control checks.
Status returned


PASS Status on GET 200 
PASS Status on GET 201 
PASS Status on GET 202 
PASS Status on GET 203 
PASS Status on GET 204 
PASS Status on GET 205 
PASS Status on GET 206 
PASS Status on GET 209 
PASS Status on GET 299 
PASS Status on POST 200 
PASS Status on HEAD 200 
PASS Status on PUT 200 
PASS Status on CHICKEN 200 
PASS Status on GET 400 
PASS Status on HEAD 401 
PASS Status on POST 404 
PASS Status on POST 500 
PASS Status on PUT 699 
PASS Status on CHICKEN 501 
PASS Status on GET 400 (nonsimple) 
PASS Status on HEAD 401 (nonsimple) 
PASS Status on POST 404 (nonsimple) 
PASS Status on PUT 699 (nonsimple) 
PASS Status on CHICKEN 501 (nonsimple) 
